Migrate chatgpt-import skill from bundled to workspace skill
Move the chatgpt-import skill from assistant/src/config/bundled-skills/ to
skills/ following the conventions in skills/AGENTS.md. The import logic
(ZIP parsing, conversation creation, message insertion, memory indexing)
is moved into a new `assistant conversations import` CLI subcommand,
replacing the TOOLS.json-based tool executor pattern.

- Add `conversations import` CLI command with --json output
- Create skills/chatgpt-import/SKILL.md referencing the CLI command
- Remove bundled skill directory and bundled-tool-registry entry
- Add chatgpt-import to skills/catalog.json
