+++ assistant/src/cli/commands/conversations.ts
@@ -1,4 +1,9 @@
+import { existsSync, readFileSync } from "node:fs";
+import { inflateRawSync } from "node:zlib";
+
 import type { Command } from "commander";
+import { eq } from "drizzle-orm";
+import { v4 as uuid } from "uuid";
 
 import {
   getQdrantUrlEnv,
@@ -12,6 +17,7 @@ import { ensureDaemonRunning } from "../../daemon/lifecycle.js";
 import { formatJson, formatMarkdown } from "../../export/formatter.js";
 import { cliIpcCall } from "../../ipc/cli-client.js";
 import {
+  addMessage,
   clearAll as clearAllConversations,
   countConversationsByScheduleJobId,
   createConversation,
@@ -20,21 +26,31 @@ import {
   wipeConversation,
 } from "../../memory/conversation-crud.js";
 import { listConversations } from "../../memory/conversation-queries.js";
+import { getDb } from "../../memory/db.js";
 import {
   selectEmbeddingBackend,
   SPARSE_EMBEDDING_VERSION,
 } from "../../memory/embedding-backend.js";
+import { indexMessageNow } from "../../memory/indexer.js";
 import { enqueueMemoryJob } from "../../memory/jobs-store.js";
 import { initQdrantClient } from "../../memory/qdrant-client.js";
+import {
+  conversationKeys,
+  conversations as conversationsTable,
+  messages as messagesTable,
+} from "../../memory/schema.js";
 import {
   initAuthSigningKey,
   loadOrCreateSigningKey,
   mintDaemonDeliveryToken,
 } from "../../runtime/auth/token-service.js";
 import { deleteSchedule } from "../../schedule/schedule-store.js";
+import { getLogger } from "../../util/logger.js";
 import { timeAgo } from "../../util/time.js";
 import { initializeDb } from "../db.js";
 import { log } from "../logger.js";
+
+const importLog = getLogger("chatgpt-import");
 
 export function registerConversationsCommand(program: Command): void {
   const conversations = program
@@ -46,13 +62,14 @@ export function registerConversationsCommand(program: Command): void {
     `
 Conversations with the assistant. Each conversation has a unique ID and a
 title. The assistant must be running for "new" (it communicates via the HTTP
-API), while "list", "export", and "clear" operate on the local SQLite database
-directly.
+API), while "list", "export", "import", and "clear" operate on the local SQLite
+database directly.
 
 Examples:
   $ assistant conversations list
   $ assistant conversations new "Project planning"
   $ assistant conversations export
+  $ assistant conversations import ~/Downloads/chatgpt-export.zip
   $ assistant conversations clear`,
   );
 
@@ -475,4 +492,410 @@ Examples:
         }
       },
     );
+
+  conversations
+    .command("import <filePath>")
+    .description(
+      "Import conversation history from a ChatGPT export ZIP archive",
+    )
+    .option("--json", "Output result as JSON")
+    .addHelpText(
+      "after",
+      `
+Arguments:
+  filePath   Absolute path to the ChatGPT export ZIP file
+
+Imports conversations from a ChatGPT data export. The ZIP file is the archive
+that ChatGPT emails after requesting a data export (Settings → Data controls →
+Export data). Only conversations.json inside the ZIP is processed.
+
+Conversations are deduplicated — re-importing the same file will skip already-
+imported conversations. Only user and assistant messages are imported (system
+prompts and tool calls are filtered out). Original timestamps from ChatGPT are
+preserved, and imported messages are indexed for memory search.
+
+Operates on the local SQLite database directly — does not require the assistant.
+
+Examples:
+  $ assistant conversations import ~/Downloads/chatgpt-export.zip
+  $ assistant conversations import /tmp/export.zip --json`,
+    )
+    .action(async (filePath: string, opts?: { json?: boolean }) => {
+      if (!filePath.endsWith(".zip")) {
+        const msg =
+          "Only ZIP files are accepted. Please provide the ChatGPT export ZIP file.";
+        if (opts?.json) {
+          log.info(JSON.stringify({ ok: false, error: msg }));
+        } else {
+          log.error(`Error: ${msg}`);
+        }
+        process.exitCode = 1;
+        return;
+      }
+
+      if (!existsSync(filePath)) {
+        const msg = `File not found: ${filePath}`;
+        if (opts?.json) {
+          log.info(JSON.stringify({ ok: false, error: msg }));
+        } else {
+          log.error(`Error: ${msg}`);
+        }
+        process.exitCode = 1;
+        return;
+      }
+
+      let imported: ChatGPTImportedConversation[];
+      try {
+        imported = parseChatGPTExport(filePath);
+      } catch (err) {
+        const msg = `Error parsing export file: ${err instanceof Error ? err.message : String(err)}`;
+        if (opts?.json) {
+          log.info(JSON.stringify({ ok: false, error: msg }));
+        } else {
+          log.error(msg);
+        }
+        process.exitCode = 1;
+        return;
+      }
+
+      if (imported.length === 0) {
+        const result = {
+          ok: true,
+          imported: 0,
+          skipped: 0,
+          messages: 0,
+        };
+        if (opts?.json) {
+          log.info(JSON.stringify(result));
+        } else {
+          log.info("No conversations found in the export file.");
+        }
+        return;
+      }
+
+      initializeDb();
+      const db = getDb();
+      let importedCount = 0;
+      let skippedCount = 0;
+      let messageCount = 0;
+
+      for (const conv of imported) {
+        const convKey = `chatgpt:${conv.sourceId}`;
+
+        const existing = db
+          .select()
+          .from(conversationKeys)
+          .where(eq(conversationKeys.conversationKey, convKey))
+          .get();
+
+        if (existing) {
+          skippedCount++;
+          continue;
+        }
+
+        const conversation = createConversation(conv.title);
+
+        for (const msg of conv.messages) {
+          await addMessage(
+            conversation.id,
+            msg.role,
+            JSON.stringify(msg.content),
+            undefined,
+            { skipIndexing: true },
+          );
+        }
+
+        // Override timestamps to match ChatGPT originals
+        db.update(conversationsTable)
+          .set({ createdAt: conv.createdAt, updatedAt: conv.updatedAt })
+          .where(eq(conversationsTable.id, conversation.id))
+          .run();
+
+        // Update message timestamps to match ChatGPT originals
+        const dbMessages = db
+          .select({ id: messagesTable.id })
+          .from(messagesTable)
+          .where(eq(messagesTable.conversationId, conversation.id))
+          .orderBy(messagesTable.createdAt)
+          .all();
+
+        const memoryConfig = getConfig().memory;
+        for (
+          let i = 0;
+          i < dbMessages.length && i < conv.messages.length;
+          i++
+        ) {
+          const originalTimestamp = conv.messages[i].createdAt;
+          db.update(messagesTable)
+            .set({ createdAt: originalTimestamp })
+            .where(eq(messagesTable.id, dbMessages[i].id))
+            .run();
+
+          try {
+            await indexMessageNow(
+              {
+                messageId: dbMessages[i].id,
+                conversationId: conversation.id,
+                role: conv.messages[i].role,
+                content: JSON.stringify(conv.messages[i].content),
+                createdAt: originalTimestamp,
+              },
+              memoryConfig,
+            );
+          } catch (err) {
+            importLog.warn(
+              "Failed to index imported message %s in conversation %s: %s",
+              dbMessages[i].id,
+              conversation.id,
+              err instanceof Error ? err.message : String(err),
+            );
+          }
+        }
+
+        db.insert(conversationKeys)
+          .values({
+            id: uuid(),
+            conversationKey: convKey,
+            conversationId: conversation.id,
+            createdAt: Date.now(),
+          })
+          .run();
+
+        importedCount++;
+        messageCount += conv.messages.length;
+      }
+
+      if (opts?.json) {
+        log.info(
+          JSON.stringify({
+            ok: true,
+            imported: importedCount,
+            skipped: skippedCount,
+            messages: messageCount,
+          }),
+        );
+      } else {
+        const lines = [
+          `Imported ${importedCount} conversation(s) with ${messageCount} message(s).`,
+        ];
+        if (skippedCount > 0) {
+          lines.push(
+            `Skipped ${skippedCount} already-imported conversation(s).`,
+          );
+        }
+        log.info(lines.join("\n"));
+      }
+    });
+}
+
+// -- ChatGPT export format types --
+
+interface ChatGPTContent {
+  content_type: string;
+  parts?: (string | null | Record<string, unknown>)[];
+}
+
+interface ChatGPTNode {
+  message: {
+    author: { role: string };
+    content: ChatGPTContent;
+    create_time?: number | null;
+  } | null;
+  parent: string | null;
+  children: string[];
+}
+
+interface ChatGPTConversation {
+  id?: string;
+  title: string;
+  create_time: number;
+  update_time: number;
+  current_node: string;
+  mapping: Record<string, ChatGPTNode>;
+}
+
+interface ChatGPTImportedMessage {
+  role: string;
+  content: Array<{ type: string; text: string }>;
+  createdAt: number;
+}
+
+interface ChatGPTImportedConversation {
+  sourceId: string;
+  title: string;
+  createdAt: number;
+  updatedAt: number;
+  messages: ChatGPTImportedMessage[];
+}
+
+// -- Parser --
+
+function parseChatGPTExport(zipPath: string): ChatGPTImportedConversation[] {
+  const jsonContent = extractConversationsJsonFromZip(zipPath);
+
+  const raw = JSON.parse(jsonContent);
+  if (!Array.isArray(raw)) {
+    throw new Error("Expected conversations.json to contain a JSON array");
+  }
+
+  const results: ChatGPTImportedConversation[] = [];
+  for (const conv of raw as ChatGPTConversation[]) {
+    const imported = parseChatGPTConversation(conv);
+    if (imported) {
+      results.push(imported);
+    }
+  }
+  return results;
+}
+
+function parseChatGPTConversation(
+  conv: ChatGPTConversation,
+): ChatGPTImportedConversation | null {
+  const { mapping, current_node } = conv;
+  if (!mapping || !current_node || !mapping[current_node]) return null;
+
+  // Walk from current_node to root via parent pointers, then reverse for chronological order
+  const nodeIds: string[] = [];
+  let nodeId: string | null = current_node;
+  while (nodeId) {
+    nodeIds.push(nodeId);
+    nodeId = mapping[nodeId]?.parent ?? null;
+  }
+  nodeIds.reverse();
+
+  const messages: ChatGPTImportedMessage[] = [];
+  for (const id of nodeIds) {
+    const node = mapping[id];
+    if (!node?.message) continue;
+
+    const { author, content, create_time } = node.message;
+    const role = author?.role;
+    if (role !== "user" && role !== "assistant") continue;
+
+    const text = extractChatGPTText(content);
+    if (!text) continue;
+
+    messages.push({
+      role,
+      content: [{ type: "text", text }],
+      createdAt: create_time
+        ? Math.round(create_time * 1000)
+        : Math.round(conv.create_time * 1000),
+    });
+  }
+
+  if (messages.length === 0) return null;
+
+  return {
+    sourceId: conv.id ?? `${conv.title}-${conv.create_time}`,
+    title: conv.title || "Untitled",
+    createdAt: Math.round(conv.create_time * 1000),
+    updatedAt: Math.round(conv.update_time * 1000),
+    messages,
+  };
+}
+
+function extractChatGPTText(content: ChatGPTContent): string {
+  if (!content?.parts) return "";
+  return content.parts
+    .filter((p): p is string => typeof p === "string")
+    .join("");
+}
+
+// -- ZIP extraction --
+
+function extractConversationsJsonFromZip(zipPath: string): string {
+  const buffer = readFileSync(zipPath);
+
+  // Find end of central directory record (EOCD signature: 0x06054b50)
+  let eocdOffset = -1;
+  for (let i = buffer.length - 22; i >= 0; i--) {
+    if (
+      buffer[i] === 0x50 &&
+      buffer[i + 1] === 0x4b &&
+      buffer[i + 2] === 0x05 &&
+      buffer[i + 3] === 0x06
+    ) {
+      eocdOffset = i;
+      break;
+    }
+  }
+  if (eocdOffset === -1) {
+    throw new Error(
+      "Invalid ZIP file: could not find end of central directory",
+    );
+  }
+
+  const centralDirOffset = buffer.readUInt32LE(eocdOffset + 16);
+  const centralDirEntries = buffer.readUInt16LE(eocdOffset + 10);
+
+  // Walk central directory to find conversations.json
+  let offset = centralDirOffset;
+  for (let i = 0; i < centralDirEntries; i++) {
+    if (
+      buffer[offset] !== 0x50 ||
+      buffer[offset + 1] !== 0x4b ||
+      buffer[offset + 2] !== 0x01 ||
+      buffer[offset + 3] !== 0x02
+    ) {
+      throw new Error("Invalid ZIP central directory entry");
+    }
+
+    const cdCompressedSize = buffer.readUInt32LE(offset + 20);
+    const fileNameLength = buffer.readUInt16LE(offset + 28);
+    const extraLength = buffer.readUInt16LE(offset + 30);
+    const commentLength = buffer.readUInt16LE(offset + 32);
+    const localHeaderOffset = buffer.readUInt32LE(offset + 42);
+    const fileName = buffer
+      .subarray(offset + 46, offset + 46 + fileNameLength)
+      .toString("utf-8");
+
+    if (
+      fileName === "conversations.json" ||
+      fileName.endsWith("/conversations.json")
+    ) {
+      return extractChatGPTLocalFile(
+        buffer,
+        localHeaderOffset,
+        cdCompressedSize,
+      );
+    }
+
+    offset += 46 + fileNameLength + extraLength + commentLength;
+  }
+
+  throw new Error("conversations.json not found in ZIP file");
+}
+
+function extractChatGPTLocalFile(
+  buffer: Buffer,
+  offset: number,
+  cdCompressedSize: number,
+): string {
+  if (
+    buffer[offset] !== 0x50 ||
+    buffer[offset + 1] !== 0x4b ||
+    buffer[offset + 2] !== 0x03 ||
+    buffer[offset + 3] !== 0x04
+  ) {
+    throw new Error("Invalid ZIP local file header");
+  }
+
+  const compressionMethod = buffer.readUInt16LE(offset + 8);
+  const localCompressedSize = buffer.readUInt32LE(offset + 18);
+  const compressedSize =
+    cdCompressedSize > 0 ? cdCompressedSize : localCompressedSize;
+  const fileNameLength = buffer.readUInt16LE(offset + 26);
+  const extraLength = buffer.readUInt16LE(offset + 28);
+
+  const dataOffset = offset + 30 + fileNameLength + extraLength;
+  const fileData = buffer.subarray(dataOffset, dataOffset + compressedSize);
+
+  if (compressionMethod === 0) {
+    return fileData.toString("utf-8");
+  } else if (compressionMethod === 8) {
+    return inflateRawSync(fileData).toString("utf-8");
+  } else {
+    throw new Error(`Unsupported ZIP compression method: ${compressionMethod}`);
+  }
 }
