--- assistant/src/config/bundled-skills/chatgpt-import/TOOLS.json
@@ -1,27 +1,0 @@
-{
-  "version": 1,
-  "tools": [
-    {
-      "name": "chatgpt_import",
-      "description": "Import ChatGPT conversation history from an exported ZIP archive. Accepts the ZIP file that ChatGPT emails after requesting a data export.",
-      "category": "import",
-      "risk": "medium",
-      "input_schema": {
-        "type": "object",
-        "properties": {
-          "file_path": {
-            "type": "string",
-            "description": "Absolute path to the ChatGPT export ZIP file"
-          },
-          "activity": {
-            "type": "string",
-            "description": "Brief non-technical explanation of why this tool is being called"
-          }
-        },
-        "required": ["file_path"]
-      },
-      "executor": "tools/chatgpt-import.ts",
-      "execution_target": "host"
-    }
-  ]
-}
