+++ skills/chatgpt-import/SKILL.md
@@ -14,14 +14,26 @@ Import ChatGPT conversation history into Vellum so users can keep their conversa
 
 When a user wants to import their ChatGPT conversations:
 
-1. **Tell the user how to export.** They need to go to ChatGPT → Settings → Data controls → Export data. ChatGPT will email them a ZIP file.
+1. **Tell the user how to export.** They need to go to ChatGPT Settings > Data controls > Export data. ChatGPT will email them a ZIP file.
 2. **Ask the user to upload the ZIP file.** Use whatever file upload tools or skills are available to receive the ZIP.
-3. **Run the import.** Once you have the ZIP file path, call `chatgpt_import` with the file path. Report the results - how many conversations and messages were imported, and any skipped duplicates.
+3. **Run the import.** Once you have the ZIP file path, run:
+
+```bash
+assistant conversations import <path-to-zip> --json
+```
+
+The command returns JSON:
+
+```json
+{ "ok": true, "imported": 12, "skipped": 0, "messages": 347 }
+```
+
+Report the results — how many conversations and messages were imported, and any skipped duplicates.
 
 ## Notes
 
 - Only ZIP files are accepted (the full export archive from ChatGPT).
-- Conversations are deduplicated - re-importing the same file will skip already-imported conversations.
+- Conversations are deduplicated — re-importing the same file will skip already-imported conversations.
 - Only user and assistant messages are imported (system prompts and tool calls are filtered out).
 - Original timestamps from ChatGPT are preserved.
 - Imported conversations are automatically indexed for memory search.
